+++ lib/Instabug.dart
@@ -6,6 +6,7 @@ import 'dart:typed_data';
 import 'dart:ui';
 
 import 'package:flutter/services.dart';
+import 'package:instabug_flutter/utils/platform_manager.dart';
 
 enum InvocationEvent {
   shake,
@@ -102,7 +103,7 @@ class Instabug {
   /// the SDK's UI.
   static Future<void> start(
       String token, List<InvocationEvent> invocationEvents) async {
-    if (Platform.isIOS) {
+    if (PlatformManager.instance.isIOS()) {
       final List<String> invocationEventsStrings =
           invocationEvents.map((e) => e.toString()).toList(growable: false);
       final List<dynamic> params = <dynamic>[token, invocationEventsStrings];